backend changes

+++ api/src/controllers/deleteDog.js
@@ -3,13 +3,16 @@ const { Dog } = require("../db");
 const deleteDog = async (req, res) => {
   try {
     const { id } = req.params;
-    await Dog.destroy({
-      where: { id },
-    });
-
-    return res.status(200).json({ message: "Dog deleted successfully" });
+    const isDBDog = typeof id === "string";
+    if (isDBDog) {
+      await Dog.destroy({
+        where: { id },
+      });
+      return res.status(200).json({ message: "Dog deleted successfully" });
+    }
+    return res.status(403).json({ error: "ApiDogs cannot be removed" });
   } catch (error) {
-    return res.status(500).json({ error: error.message });
+    return res.status(500).json({ error: "Failed to delete dog" });
   }
 };
